+++ llvm/lib/Transforms/HLSL/DXILFinalizeLinkage.cpp
@@ -6,8 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "DXILFinalizeLinkage.h"
-#include "DirectX.h"
+#include "llvm/Transforms/HLSL/DXILFinalizeLinkage.h"
 #include "llvm/IR/Function.h"
 #include "llvm/IR/GlobalValue.h"
 #include "llvm/IR/Metadata.h"
@@ -44,19 +43,4 @@ PreservedAnalyses DXILFinalizeLinkage::run(Module &M,
   if (finalizeLinkage(M))
     return PreservedAnalyses::none();
   return PreservedAnalyses::all();
-}
-
-bool DXILFinalizeLinkageLegacy::runOnModule(Module &M) {
-  return finalizeLinkage(M);
-}
-
-char DXILFinalizeLinkageLegacy::ID = 0;
-
-INITIALIZE_PASS_BEGIN(DXILFinalizeLinkageLegacy, DEBUG_TYPE,
-                      "DXIL Finalize Linkage", false, false)
-INITIALIZE_PASS_END(DXILFinalizeLinkageLegacy, DEBUG_TYPE,
-                    "DXIL Finalize Linkage", false, false)
-
-ModulePass *llvm::createDXILFinalizeLinkageLegacyPass() {
-  return new DXILFinalizeLinkageLegacy();
 }
